A document

A document

documento(cve)

Arguments

cve
A character with a document CVE.

See also

Examples

document_csv <- "BOCG_D_14_110_901"
documento(document_csv)

#>   SECC NBOL ANNO DISP       INIC
#> 1    S  110 2020  901 661/000073
#>                                                                                                                                                                                                                                                                                                                                                                                                                        TITULO
#> 1 Del GRUPO PARLAMENTARIO IZQUIERDA CONFEDERAL (ADELANTE ANDALUCÍA, MÉS PER MALLORCA, MÁS MADRID, COMPROMÍS, GEROA BAI Y CATALUNYA EN COMÚ PODEM), por la que se insta al Gobierno a la modificación y actualización de la Ley 52/2007, de 26 de diciembre, por la que se reconocen y amplían derechos y se establecen medidas en favor de quienes padecieron persecución o violencia durante la guerra civil y la dictadura.
#>                     FASE      FECHA               CVE     apartado
#> 1 Acuerdo de la Comisión 2020-11-20 BOCG_D_14_110_901 II. Mociones
#>            subapartado numpag numpagfin DISP     NUMEXP
#> 1 Mociones en Comisión      3         4  901 661/000073
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                       text
#> 1 La Mesa de la Cámara, en su reunión del día 17 de noviembre de 2020, ha adoptado el acuerdo que se indica respecto al asunto de referencia:Moción en Comisión.661/000073AUTOR: COMISIÓN CONSTITUCIONALComunicación del acuerdo adoptado por la Comisión Constitucional, en su sesión del día 27 de octubre de 2020, por el que se rechaza la moción suscrita por el Grupo Parlamentario Izquierda Confederal (Adelante Andalucía, Més per Mallorca, Más Madrid, Compromís, Geroa Bai y Catalunya en Comú Podem), por la que se insta al Gobierno a la modificación y actualización de la Ley 52/2007, de 26 de diciembre, por la que se reconocen y amplían derechos y se establecen medidas en favor de quienes padecieron persecución o violencia durante la guerra civil y la dictadura, con la incorporación de las dos enmiendas presentadas por el Grupo Parlamentario Esquerra Republicana-Euskal Herria Bildu.ACUERDO:Tomar conocimiento y publicar en el Boletín Oficial de las Cortes Generales.En ejecución de dicho acuerdo, se ordena la publicación de conformidad con el artículo 191 del Reglamento de la Cámara.Palacio del Senado, 17 de noviembre de 2020.—P.D., Manuel Cavero Gómez, Letrado Mayor del Senado.La Comisión Constitucional del Senado, en su sesión celebrada el día 27 de octubre de 2020, ha rechazado la moción del Grupo Parlamentario Izquierda Confederal (Adelante Andalucía, Més per Mallorca, Más Madrid, Compromís, Geroa Bai y Catalunya en Comú Podem) por la que se insta al Gobierno a la modificación y actualización de la Ley 52/2007, de 26 de diciembre, por la que se reconocen y amplían derechos y se establecen medidas en favor de quienes padecieron persecución o violencia durante la guerra civil y la dictadura (661/000073), con la incorporación de las dos enmiendas presentadas por el Grupo Parlamentario Esquerra Republicana-Euskal Herria Bildu, con el siguiente texto:«La Comisión Constitucional del Senado insta al Gobierno a:1. Modificar la Ley 52/2007, de 26 de diciembre, por la que se reconocen y amplían derechos y se establecen medidas en favor de quienes padecieron persecución o violencia durante la guerra civil y la dictadura, con el objetivo de reconocer que la exhumación y apertura de fosas comunes es una tarea de responsabilidad pública, y por lo tanto son las Administraciones Públicas quienes tienen la obligación y el deber de acometer esa tarea.2. Actualizar la Ley 52/2007, y a seguir el camino de las CC. AA. en diversos aspectos recogidos en las leyes y decretos autonómicos sobre Memoria Histórica.3. Crear un  fondo finalista, con cargo a los Presupuestos Generales del Estado, destinado a sufragar los costes derivados de la exhumación y apertura de fosas comunes por las Administraciones Públicas.»