Organization chart

Retrieves the relationships between people responsible of the Senate. At the moment due to limitation on the website only works on the 13 legislature.

organigrama(legislatura)

Arguments

legislatura
A numeric value above 12 (No information prior to then).

Value

A data.frame with all the information available.

Examples

head(organigrama(13))

#>   nivel
#> 1     1
#> 2     2
#> 3     3
#> 4     3
#> 5     2
#> 6     3
#>                                                              nombreOficial
#> 1                                                       Secretaría General
#> 2                                          Gabinete del Secretario General
#> 3 Gabinete dependiente de la Secretaría General de Unidad de asesoramiento
#> 4                Unidad de Coordinación, control y evaluación de proyectos
#> 5                                        Servicio de Dotación de Seguridad
#> 6                                         Unidad de Gestión Administrativa
#>                                                      nombre codigo codigoPadre
#> 1                                        Secretaría General      8            
#> 2                           Gabinete del Secretario General    223           8
#> 3                                   Unidad de asesoramiento    218         223
#> 4 Unidad de Coordinación, control y evaluación de proyectos    207         223
#> 5                                     Dotación de Seguridad    208           8
#> 6                                    Gestión Administrativa    235         208
#>   tipoDependencia                jefe                        dirCorreo
#> 1              AD MANUEL CAVERO GÓMEZ secretarialetradomayor@senado.es
#> 2             GAB                                                     
#> 3             GAB                                                     
#> 4             GAB                                                     
#> 5             SER                                                     
#> 6             UNI                                                     
#>    numTelefono
#> 1 91 538 14 00
#> 2             
#> 3             
#> 4             
#> 5             
#> 6